GROOVY-9742: GroovyClassLoader.parseClass() StampedCommonCache.getAndPut() hang (provide constructor to enable alternative sourceCaches to be set)
